With wifidog

+++ src/main/java/com/acs_tr069/test_tr069/radius/entity/Subscribers.java
@@ -31,7 +31,7 @@ public class Subscribers {
     private Integer sessionLimit;
 
     @Column(name = "remaining_session_time")
-    private Long remainingSessionTime;
+    private Integer remainingSessionTime;
 
     @Column(name = "bytes_limit")
     private Long bytesLimit;
